Extracting BfPipelineConfig methods to utility class
