+++ bazel/deps.bzl
@@ -175,6 +175,20 @@ def stratum_deps():
             build_file = "@//bazel:external/yang.BUILD",
         )
 
+    if "com_github_libarchive_libarchive" not in native.existing_rules():
+        http_archive(
+            name = "com_github_libarchive_libarchive",
+            url = "https://github.com/libarchive/libarchive/releases/download/v3.4.3/libarchive-3.4.3.tar.gz",
+            strip_prefix = "libarchive-3.4.3",
+            sha256 = "ee1e749213c108cb60d53147f18c31a73d6717d7e3d2481c157e1b34c881ea39",
+            patch_cmds = [
+                "./configure --prefix $(realpath ./local-install/) --enable-shared=no --without-openssl",
+                "make -j",
+                "make install",
+            ],
+            build_file = "@//bazel:external/libarchive.BUILD",
+        )
+
     if "com_github_nlohmann_json" not in native.existing_rules():
         http_archive(
             name = "com_github_nlohmann_json",
